# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Meeting Slot Service API
  version: 1.0.0
  description: |
    REST API for managing meeting slots and finding optimal meeting times for distributed teams.
    
    This service helps teams across multiple time zones find the best available meeting slots
    based on participant availability, working hours, and configurable preferences.
servers:
  - url: https://{api-gateway-id}.execute-api.{region}.amazonaws.com/v1
    description: AWS API Gateway
    variables:
      api-gateway-id:
        default: xxxxx
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS Region
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/events/{id}/participants/{user_id}/availability:
    post:
      tags:
        - Availability
      summary: Submit availability
      description: Submits participant's availability for an event
      operationId: submitAvailability
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
        - $ref: '#/components/parameters/UserIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
            example:
              available_slots:
                - start_time: "2026-02-01T14:00:00+05:30"
                  end_time: "2026-02-01T16:00:00+05:30"
                  timezone: "Asia/Kolkata"
                - start_time: "2026-02-01T17:00:00+05:30"
                  end_time: "2026-02-01T19:00:00+05:30"
                  timezone: "Asia/Kolkata"
      responses:
        '200':
          description: Availability submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                data:
                  message: "Availability submitted successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BAD_REQUEST"
                  message: "participant not found"

    put:
      tags:
        - Availability
      summary: Update availability
      description: Updates participant's availability for an event
      operationId: updateAvailability
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
        - $ref: '#/components/parameters/UserIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
      responses:
        '200':
          description: Availability updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                success: true
                data:
                  message: "Availability updated successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Availability
      summary: Get participant availability
      description: Retrieves participant's availability for an event
      operationId: getAvailability
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
        - $ref: '#/components/parameters/UserIdPathParam'
      responses:
        '200':
          description: Participant availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityListResponse'
              example:
                success: true
                data:
                  - id: 1
                    event_id: "evt_xyz789"
                    user_id: "usr_def456"
                    start_time: "2026-02-01T14:00:00+05:30"
                    end_time: "2026-02-01T16:00:00+05:30"
                    timezone: "Asia/Kolkata"
                    created_at: "2026-02-18T12:00:00Z"
                    updated_at: "2026-02-18T12:00:00Z"
                  - id: 2
                    event_id: "evt_xyz789"
                    user_id: "usr_def456"
                    start_time: "2026-02-01T17:00:00+05:30"
                    end_time: "2026-02-01T19:00:00+05:30"
                    timezone: "Asia/Kolkata"
                    created_at: "2026-02-18T12:00:00Z"
                    updated_at: "2026-02-18T12:00:00Z"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Availability
      summary: Delete availability
      description: Deletes participant's availability for an event
      operationId: deleteAvailability
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
        - $ref: '#/components/parameters/UserIdPathParam'
      responses:
        '204':
          description: Availability deleted successfully
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}/recommendations:
    get:
      tags:
        - Recommendations
      summary: Get meeting slot recommendations
      description: Calculates and returns the best meeting slot recommendation based on participant availability
      operationId: getRecommendations
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: Recommendation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
              example:
                success: true
                data:
                  event_id: "evt_xyz789"
                  duration_minutes: 60
                  total_participants: 4
                  best_recommendation:
                    slot:
                      start_time: "2026-02-01T15:00:00+05:30"
                      end_time: "2026-02-01T16:00:00+05:30"
                      timezone: "Asia/Kolkata"
                    available_participants: 4
                    availability_rate: 1.0
                    available_users: ["usr_def456", "usr_ghi789", "usr_jkl012", "usr_mno345"]
                    unavailable_users: []
                  message: "Perfect match! All 4 participants are available for this time slot"
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: No recommendations available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BAD_REQUEST"
                  message: "no overlapping availability found"
components:
  parameters:
...
    UserIdPathParam:
      name: user_id
      in: path
      required: true
      description: User unique identifier
      schema:
        type: string
        example: "usr_def456"
    EventIdParam:
      name: id
      in: path
      required: true
      description: Event unique identifier
      schema:
        type: string
        example: "evt_xyz789"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        message:
          type: string
          description: Error message
          example: "Resource not found"
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "Operation completed successfully"
...
    TimeSlot:
      type: object
      required:
        - start_time
        - end_time
        - timezone
      properties:
        start_time:
          type: string
          format: date-time
          description: Slot start time
          example: "2026-02-20T09:00:00Z"
        end_time:
          type: string
          format: date-time
          description: Slot end time
          example: "2026-02-20T10:00:00Z"
        timezone:
          type: string
          description: Timezone
          example: "UTC"

    # User Schemas
...
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
          description: Slot ID
          example: 1
        event_id:
          type: string
          description: Associated event ID
          example: "evt_xyz789"
        user_id:
          type: string
          description: Associated user ID
          example: "usr_def456"
        start_time:
          type: string
          format: date-time
          description: Availability start time
          example: "2026-02-01T14:00:00+05:30"
        end_time:
          type: string
          format: date-time
          description: Availability end time
          example: "2026-02-01T16:00:00+05:30"
        timezone:
          type: string
          description: Timezone
          example: "Asia/Kolkata"
        created_at:
          type: string
          format: date-time
          example: "2026-02-18T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-02-18T12:00:00Z"
    AvailabilityRequest:
      type: object
      required:
        - available_slots
      properties:
        available_slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
    AvailabilityListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilitySlot'

    # Recommendation Schemas
    RecommendationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            event_id:
              type: string
              description: Event ID
              example: "evt_xyz789"
            duration_minutes:
              type: integer
              description: Event duration
              example: 60
            total_participants:
              type: integer
              description: Total number of participants
              example: 4
            best_recommendation:
              $ref: '#/components/schemas/Recommendation'
            message:
              type: string
              description: Human-readable message about the recommendation
              example: "Perfect match! All 4 participants are available for this time slot"
    Recommendation:
      type: object
      properties:
        slot:
          $ref: '#/components/schemas/TimeSlot'
        available_participants:
          type: integer
          description: Number of participants available
          example: 4
        availability_rate:
          type: number
          format: float
          description: Availability rate (0.0 to 1.0)
          example: 1.0
        available_users:
          type: array
          items:
            type: string
          description: List of available user IDs
          example: ["usr_def456", "usr_ghi789", "usr_jkl012", "usr_mno345"]
        unavailable_users:
          type: array
          items:
            type: string
          description: List of unavailable user IDs
          example: []